feat: added a feature to guess the date of the riddle (#45)
* By default, the application will try to guess the year
  and the day user meant when using `elv` based on the current
  date and the arguments provided by the user.
* Brushed up on the documentation.

Closes #38

+++ src/domain/ports/aoc_client.rs
@@ -6,8 +6,8 @@ use super::errors::AocClientError;
 
 pub trait AocClient {
     fn submit_answer(&self, submission: Submission) -> Result<SubmissionResult, AocClientError>;
-    fn get_description<Desc>(&self, year: &u16, day: &u8) -> Result<Desc, AocClientError>
+    fn get_description<Desc>(&self, year: &i32, day: &i32) -> Result<Desc, AocClientError>
     where
         Desc: Description + TryFrom<reqwest::blocking::Response>;
-    fn get_input(&self, year: &u16, day: &u8) -> InputResponse;
+    fn get_input(&self, year: &i32, day: &i32) -> InputResponse;
 }
